requirements.txt
discord.py
httpx
openai
pyyaml
schedule